+++ src/lib/components/Dark/DarkProvider.svelte
@@ -11,7 +11,7 @@
 	<script>
 		// to prevent white screen on page load
 		// until the dark mode is initialized from the store
-		const isDarkMode = !!localStorage.getItem('scheme-dark');
+		const isDarkMode = !!localStorage.getItem("hds-dark");
 		if (isDarkMode) {
 			document.documentElement.classList.add('dark');
 		}