+++ src/routes/(docs)/[[slug]]/+page.svelte
@@ -19,6 +19,7 @@
 		{ href: '/avatar', label: 'Avatar' },
 		{ href: '/button', label: 'Button' },
 		{ href: '/box', label: 'Box' },
+		{ href: '/box-shadow-picker', label: 'Box Shadow Picker' },
 		{ href: '/color-picker', label: 'Color Picker' },
 		{ href: '/icon-button', label: 'Icon Button' },
 		{ href: '/callout', label: 'Callout' },
@@ -41,6 +42,7 @@
 		{ href: '/tag', label: 'Tag' },
 		{ href: '/tooltip', label: 'Tooltip' },
 		{ href: '/toast', label: 'Toast' },
+		{ href: '/usage', label: 'Usage' },
 		{ href: '/modal', label: 'Modal' },
 		{ href: '/icon-message', label: 'Icon Message' },
 		{ href: '/slider', label: 'Slider' }